Updated base.html, Implemented a ticker profile, Created landing
So now the project has a lot prettier things. and the start of the profiles feature which will be the first step of any playbook.  so now it's time to focus on fleshing out the report card, and build out the FE JS to prepare the factor section, implement users and auth.